requirements.txt
ImageHash==4.3.1
pandas==2.2.3
Pillow==10.4.0
PyExifTool==0.5.6
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
